Use the fixed-width spellings for integers
227 declarations: unsigned long -> uint32_t (118), unsigned short ->
uint16_t (69), unsigned int -> uint16_t (40).  Same types, so all seven
targets are byte-identical; the spelling now states the width instead of
leaving the reader to know that int is 16 bits here.

One site cannot follow.  short and int are both two bytes, but they are
distinct types, and pointer compatibility goes by type rather than by
representation -- uint16_t is unsigned int here, so uint16_t* is not
unsigned short*.  freeze_sprited.c's mx and my have their address taken
by mouse_update_position(unsigned short*, unsigned short*), so they keep
libc's spelling, with a comment saying why.  The compiler proved that is
the only one: -Wincompatible-pointer-types is an error here.

unsigned char is left alone.  Unlike the others it is ambiguous in this
tree, where freeze_menu[] and friends hold text in it.

+++ src/cc65compat.c
@@ -4,13 +4,13 @@
 
 char* itoa(int value, char* s, int radix) {
     char* p = s;
-    unsigned int u;
+    uint16_t u;
 
     if (radix == 10 && value < 0) {
         *p++ = '-';
-        u = (unsigned int)-value;
+        u = (uint16_t)-value;
     } else {
-        u = (unsigned int)value;
+        u = (uint16_t)value;
     }
 
     utoa(u, p, radix);
