+++ src/monitor/fdisk_screen_monitor.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <stdint.h>
+
 /* MEGA65 only: writes the VIC-IV screen directly. */
 
 void setup_screen(void);
@@ -13,5 +15,5 @@ void set_screen_attributes(long screen_address, unsigned char count, unsigned ch
 char read_line(char* buffer, unsigned char max_length, unsigned char column);
 void write_prompt(const char* text, char length);
 void next_line(void);
-void fatal_error(const unsigned char* filename, unsigned int line_number);
+void fatal_error(const unsigned char* filename, uint16_t line_number);
 #define FATAL_ERROR fatal_error(__FILE__, __LINE__)
